MAINT: Updating readme
This is the final patch to indicate what changes have been made to the
core functionality of the plugin, all basic functionality has been
implemented, the rest of this plugin focusses on customisation and
extending the functionality to have more accessable features. (See
README to indicate the future goals of this plugin).

Closes #2
Closes #1